# Migrate enhanced-resolve to oxc-resolver