--- a/packages/knip/package.json
+++ b/packages/knip/package.json
@@ -62,12 +62,12 @@
   ],
   "dependencies": {
     "@nodelib/fs.walk": "^1.2.3",
-    "enhanced-resolve": "^5.18.1",
     "fast-glob": "^3.3.3",
     "formatly": "^0.2.3",
     "jiti": "^2.4.2",
     "js-yaml": "^4.1.0",
     "minimist": "^1.2.8",
+    "oxc-resolver": "^9.0.2",
     "picocolors": "^1.1.0",
     "picomatch": "^4.0.1",
     "smol-toml": "^1.3.1",