##  Dependency Update

* Removed `enhanced-resolve`
* Added `oxc-resolver`

## Code changes

* Adjust `util/resolve` code for `oxc-resolver`

## Tests

- All `yarn test` suits successful
- All `yarn test:smoke` suits successful

## Performance improvements

Haven't benchmarked this specifically, but I see about **50%** time improvement in the projects that I'm currently working on.