@agneym 2025-05-11
I see another PR #1079 here with an OXC fork used on rspack as well. Worth checking that out as well, don't mind closing this one.