@webpro 2025-05-13
Thanks for the PR, Agney!

This change improves `resolveSync` performance significantly. I've ran it a few times on two repos I already had cloned on my machine:

## [Astro](https://github.com/withastro/astro)

Before

```
Name         size  min   max   median   sum  
-----------  ----  ----  ----  ------  ------
resolveSync  2192  0.02  6.26    0.08  296.10
resolveSync  2192  0.01  9.99    0.09  315.74
resolveSync  2192  0.01  9.98    0.09  310.10
```

After

```
Name         size  min   max   median   sum 
-----------  ----  ----  ----  ------  -----
resolveSync  2192  0.00  3.17    0.02  56.00
resolveSync  2192  0.00  0.41    0.02  54.34
resolveSync  2192  0.00  0.42    0.02  54.30
```

## [Storybook](https://github.com/storybookjs/storybook)

Before

```
Name         size  min    max   median   sum  
-----------  ----  ----  -----  ------  ------
resolveSync  7600  0.01   8.72    0.05  653.81
resolveSync  7600  0.01  16.24    0.05  631.81
resolveSync  7600  0.01  16.03    0.05  639.96
resolveSync  7600  0.01  15.08    0.05  694.67
```

After

```
Name         size  min   max   median   sum  
-----------  ----  ----  ----  ------  ------
resolveSync  7600  0.00  1.65    0.01  133.84
resolveSync  7600  0.00  0.48    0.01  106.95
resolveSync  7600  0.00  6.68    0.01  113.73
```

I've just pushed a change to `main`:

- Includes `timerify` on `resolveSync` when used from `createCustomModuleResolver`
- Run `knip --performance-fn resolveSync` so we `timerify` and focus on only one function, for less performance overhead of `timerify` itself and a tiny table in the terminal output

Would love to merge either this or #1079 if you could make CI go green and get some numbers from your project(s) too.