@webpro 2025-05-13
By the way would be great if we could somehow get an idea of the overhead of loading the bindings (see `index.js` in https://www.npmjs.com/package/oxc-resolver/v/1.6.0?activeTab=code). I mean, that probably won't take half a second but also I don't have an idea.