@agneym 2025-05-13
@webpro 

- [x] Rebased 
- [x] CI should be green now 

Numbers from a project I'm running:

**Before**
---
createProgram                       1  8617.94  8617.94  8617.94  8617.94
resolveModuleNames              57067     0.00    16.02     0.00  5939.00
resolveSync                     37773     0.04     3.64     0.11  4593.64

**After**
---
createProgram                       1  4307.32  4307.32  4307.32  4307.32
resolveModuleNames              57067     0.00     7.98     0.00  1624.02
resolveSync                     37719     0.00     2.12     0.00   376.78

In percentages:

| Function             | |
|----------------------|:--------:|
| `resolveSync`        |   92%    |
| `resolveModuleNames` |   73%    |
| `createProgram`      |   50%    |