@webpro 2025-05-13
Thank you!

I meant to use literally `knip --performance-fn resolveSync` to measure only `resolveSync`, but the picture is very clear! A >90% and 4 second improvement for this function alone is simply massive 🚀