@webpro 2025-05-14
:rocket: _This pull request is included in v5.56.0. See [Release 5.56.0](https://github.com/webpro-nl/knip/releases/tag/5.56.0) for release notes._

_Using Knip in a commercial project? Please consider [becoming a sponsor](https://knip.dev/sponsors)._