@Boshen 2025-05-16
> I see another PR https://github.com/webpro-nl/knip/pull/1079 here with an OXC fork used on rspack as well. Worth checking that out as well, don't mind closing this one.

I haven't enabled Yarn PnP support yet, I'll take a look at enabling it if there are requirements.