@webpro 2025-05-16
@Boshen Knip itself doesn't actively support PnP (related: #924).